requirements.txt
torch
torchvision
pytorch-lightning
segmentation-models-pytorch
Pillow
opencv-python
wandb
tqdm
jupyterlab
matplotlib
